Fix extra trailing comma in sass selectors
Nice package! I am working on a ui-library with nestable components that need to be aware of their parent's inset padding to adjust border-radius values accordingly and was trying to jerry-rig something of my own, but this does the trick wonderfully! Using sass, I ran into an error due to an extra comma in the generated selectors, this PR fixes it.